feat(gauge): Add data index and series index for indexing the data by element
