+++ src/data/List.ts
@@ -39,7 +39,7 @@ import {
     DecalObject
 } from '../util/types';
 import {isDataItemOption, convertOptionIdName} from '../util/model';
-import { getECData } from '../util/innerStore';
+import {getECData, setCommonECData} from '../util/innerStore';
 import type Graph from './Graph';
 import type Tree from './Tree';
 import type { VisualMeta } from '../component/visualMap/VisualMapModel';
@@ -180,7 +180,6 @@ let normalizeDimensions: (dimensions: ItrParamDims) => Array<DimensionLoose>;
 let validateDimensions: (list: List, dims: DimensionName[]) => void;
 let cloneListForMapAndSample: (original: List, excludeDimensions: DimensionName[]) => List;
 let getInitialExtent: () => [number, number];
-let setItemDataAndSeriesIndex: (this: Element, child: Element) => void;
 let transferProperties: (target: List, source: List) => void;
 
 
@@ -1905,21 +1904,10 @@ class List<
      * Set graphic element relative to data. It can be set as null
      */
     setItemGraphicEl(idx: number, el: Element): void {
-        const hostModel = this.hostModel;
-
-        if (el) {
-            const ecData = getECData(el);
-            // Add data index and series index for indexing the data by element
-            // Useful in tooltip
-            ecData.dataIndex = idx;
-            ecData.dataType = this.dataType;
-            ecData.seriesIndex = hostModel && (hostModel as any).seriesIndex;
-
-            // TODO: not store dataIndex on children.
-            if (el.type === 'group') {
-                el.traverse(setItemDataAndSeriesIndex, el);
-            }
-        }
+        const seriesIndex = this.hostModel && (this.hostModel as any).seriesIndex;
+        // Add data index and series index for indexing the data by element
+        // Useful in tooltip
+        setCommonECData(seriesIndex, this.dataType, idx, el);
 
         this._graphicEls[idx] = el;
     }
@@ -2211,14 +2199,6 @@ class List<
 
         getInitialExtent = function (): [number, number] {
             return [Infinity, -Infinity];
-        };
-
-        setItemDataAndSeriesIndex = function (this: Element, child: Element): void {
-            const childECData = getECData(child);
-            const thisECData = getECData(this);
-            childECData.seriesIndex = thisECData.seriesIndex;
-            childECData.dataIndex = thisECData.dataIndex;
-            childECData.dataType = thisECData.dataType;
         };
 
         transferProperties = function (target: List, source: List): void {
